[Network] Fixed IOCP accept recovery and socket handle width
* Re-armed AcceptEx after failed IOCP accept completions or accepted socket setup failures

* Preserved pointer-sized Winsock handles through IOCP, SocketHandle and external handle assignment

+++ src/network/win32/iocp.h
@@ -6,7 +6,7 @@
 
 #include <kotuku/system/errors.h>
 
-typedef uint32_t WSW_SOCKET;
+typedef uintptr_t WSW_SOCKET;
 typedef ERR (*iocp_post_message)(int MsgID, const void *Message, int Size);
 
 static constexpr size_t IOCP_ENDPOINT_STORAGE_SIZE = 128;
